Fixed issues assigned to RawanAlShami
- Course.java : fixed imports users.Professor , users.Student  and users.TA
- Professor.java: fixed package name to users and parameter name to phDTopic to match naming conventions
- Student.java : fixed package name to users and parameter name to id to match naming conventions
- TA.java : fixed package name to users
