+++ Notifications/bin/Messages/TaskAddedEmailMessage.class
