+++ Notifications/src/messages/DailyNewsEmailMessage.java
@@ -3,6 +3,7 @@ package messages;
 public class DailyNewsEmailMessage {
 	
 	public String prepareMessage() {
+		// code to replace place holders of this type
 		return "";
 	}
 }
