+++ Notifications/src/messages/DailyNewsMobileMessage.java
@@ -1,8 +1,8 @@
-package Messages;
+package messages;
 
 public class DailyNewsMobileMessage {
 	
-	public String prepareMessage(String placeHolders[]) {
+	public String prepareMessage(String []placeHolders) {
 		// code to replace place holders of this type
 		return "";
 	}
