+++ Notifications/src/messages/GradesAnnouncementEmailMessage.java
@@ -1,8 +1,8 @@
-package Messages;
+package messages;
 
-public class GradesAnnouncementMobileMessage {
+public class GradesAnnouncementEmailMessage {
 
-	public String prepareMessage(String placeHolders[]) {
+	public String prepareMessage(String []placeHolders) {
 		// code to replace place holders of this type
 		return "";
 	}
