+++ Notifications/src/messages/GradesAnnouncementMobileMessage.java
@@ -1,8 +1,8 @@
-package Messages;
+package messages;
 
-public class GradesAnnouncementEmailMessage {
+public class GradesAnnouncementMobileMessage {
 
-	public String prepareMessage(String placeHolders[]) {
+	public String prepareMessage(String []placeHolders) {
 		// code to replace place holders of this type
 		return "";
 	}
