+++ Notifications/src/messages/TaskAddedEmailMessage.java
@@ -1,8 +1,8 @@
-package Messages;
+package messages;
 
 public class TaskAddedEmailMessage {
 
-	public String prepareMessage(String placeHolders[]) {
+	public String prepareMessage(String []placeHolders) {
 		// code to replace place holders of this type
 		
 		return "";
@@ -10,6 +10,7 @@ public class TaskAddedEmailMessage {
 	
 	
 	public void addTeamDescription() {
+		//DO NOTHING BECAUSE THIS CODE IS SKELETAL
 		
 	}
 }
