+++ Notifications/src/messages/TaskAddedMobileMessage.java
@@ -1,14 +1,14 @@
-package Messages;
+package messages;
 
 public class TaskAddedMobileMessage {
 
-	public String prepareMessage(String placeHolders[]) {
+	public String prepareMessage(String []placeHolders) {
 		// code to replace place holders of this type
 		return "";
 	}
 	
 	
 	public void addTeamDescription() {
-		
+		//DO NOTHING BECAUSE THIS CODE IS SKELETAl
 	}
 }
