+++ Notifications/src/users/TA.java
@@ -1,28 +1,22 @@
-package Users;
+package users;
 
-import java.sql.Date;
-
-public class Professor {
+public class TA {
 	String name;
 	String department;
-	Date hirringDate;
-	String PhDTopic;
 	String email;
 	String phoneNumber;
 	
 	
-	public Professor(String name, String department, Date hirringDate, String phDTopic, String email,
-			String phoneNumber) {
+
+	public TA(String name, String department, String email, String phoneNumber) {
 		super();
 		this.name = name;
 		this.department = department;
-		this.hirringDate = hirringDate;
-		PhDTopic = phDTopic;
 		this.email = email;
 		this.phoneNumber = phoneNumber;
 	}
-	
-	
+
+
 	public String getEmail() {
 		return email;
 	}
@@ -42,38 +36,25 @@ public class Professor {
 		this.phoneNumber = phoneNumber;
 	}
 
-
 	public String getName() {
 		return name;
 	}
+
 	public void setName(String name) {
 		this.name = name;
 	}
+
 	public String getDepartment() {
 		return department;
 	}
+
 	public void setDepartment(String department) {
 		this.department = department;
 	}
-	public Date getHirringDate() {
-		return hirringDate;
-	}
-	public void setHirringDate(Date hirringDate) {
-		this.hirringDate = hirringDate;
-	}
-	public String getPhDTopic() {
-		return PhDTopic;
-	}
-	public void setPhDTopic(String phDTopic) {
-		PhDTopic = phDTopic;
-	}
 	
-	
-	public void notifyProfessor(String message) {
+	public void notifyTA(String message) {
 		// do some stuff
 		
 	}
-	
-	
 	
 }
